Invite SDK UG_First Half_Rough Draft
This document represents a walkthrough update to only the first half of
the document. I do have a couple of questions to start with:

1) I’ve changed the term “App key” to “Branch Key.” I’ve seen the use
of these terms interchangeably—is that correct—do they mean the same?
Or are they two terms with different meanings? If they are the same, do
we prefer one term over the other? (I want to be consistent with our
nomenclature). I’m using Branch Key throughout the doc.

2) Along those same lines, is ‘Registering the Key’ the same as
creating an account/sign up?

Notes:

- For now, I removed the Invite Flow and the New User Experience, but
I’ll replace it towards the end of the document.

- Also, I would like to update the Introduction section. I’ll try to
update it, and then I’ll ask for feedback.

Thanks!